+++ JS/comments.js
@@ -1,0 +1,18 @@
+export let comments = [
+    // {
+    //   userName: "Глеб Фокин",
+    //   commentDate: "13.02.22 19:22",
+    //   userComment: "Это будет первый комментарий на этой странице",
+    //   likes: 3,
+    //   isLiked: false,
+    //   isEdit: false,
+    // },
+    // {
+    //   userName: "Варвара Н.",
+    //   commentDate: "13.02.22 19:22",
+    //   userComment: "Мне нравится как оформлена эта страница! ❤",
+    //   likes: 75,
+    //   isLiked: false,
+    //   isEdit: false,
+    // },
+];
