+++ JS/main.js
@@ -1,4 +1,10 @@
 "use strict";
+import { comments } from "./comments.js";
+import { renderComments } from "./renderComments.js";
+import { getFromApiFirstTime, getFromApi, postToApi } from "./api.js";
+import getCommentsList from "./CommentsList.js"
+
+
 const listElem = document.getElementById('list-comments');
 const commentBtn = document.getElementById('form-add-button');
 const addName = document.getElementById('add-name');
@@ -9,76 +15,60 @@ const loadingComments = document.querySelector('.loading');
 const commentAddedElem = document.querySelector('.comment-added');
 const addFormElem = document.querySelector('.add-form');
 
-let comments = [
-    // {
-    //   userName: "Глеб Фокин",
-    //   commentDate: "13.02.22 19:22",
-    //   userComment: "Это будет первый комментарий на этой странице",
-    //   likes: 3,
-    //   isLiked: false,
-    //   isEdit: false,
-    // },
-    // {
-    //   userName: "Варвара Н.",
-    //   commentDate: "13.02.22 19:22",
-    //   userComment: "Мне нравится как оформлена эта страница! ❤",
-    //   likes: 75,
-    //   isLiked: false,
-    //   isEdit: false,
-    // },
-];
 
-getFromApiFirstTime();
+getFromApiFirstTime(comments, loadingComments, renderComments);
+console.log(getFromApiFirstTime(comments, renderComments));
 
 
-function getFromApiFirstTime() {
-    loadingComments.style.display = 'block';
+// function getFromApiFirstTime(data, render) {
+//     loadingComments.style.display = 'block';
+//     return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
+//         method: "GET",
+//     }).then((responce) => {
+//         loadingComments.style.display = 'none';
+//         if (responce.status === 500) {
+//             throw new Error("Ошибка 500");
+//         } else {
+//             return responce.json();
+//         }
+//     }).then((responceData) => {
+//         data = responceData.comments;
+//         console.log(data);
+//         renderComments(listElem, getCommentsList);
+//         console.log(renderComments(listElem, getCommentsList));
+//     }).catch((error) => {
+//         if (error.message === "Ошибка 500") {
+//             console.log(error);
+//             console.log("Сервер сломался, попробуй позже");
+//         } else {
+//             console.log(error);
+//             console.log("Кажется, у вас сломался интернет, попробуйте позже");
+//         };
+//     })
+// };
 
-    return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
-        method: "GET",
-    }).then((responce) => {
-        loadingComments.style.display = 'none';
 
-        if (responce.status === 500) {
-            throw new Error("Ошибка 500");
-        } else {
-            return responce.json();
-        }
-    }).then((responceData) => {
-        comments = responceData.comments;
-        renderComments();
-    }).catch((error) => {
-        if (error.message === "Ошибка 500") {
-            console.log(error);
-            alert("Сервер сломался, попробуй позже");
-        } else {
-            alert("Кажется, у вас сломался интернет, попробуйте позже");
-        };
-    })
-};
-
-
-function getFromApi() {
-    return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
-        method: "GET",
-    }).then((responce) => {
-        if (responce.status === 500) {
-            throw new Error("Ошибка 500");
-        } else {
-            return responce.json();
-        }
-    }).then((responceData) => {
-        comments = responceData.comments;
-        renderComments();
-    }).catch((error) => {
-        if (error.message === "Ошибка 500") {
-            console.log(error);
-            alert("Сервер сломался, попробуй позже");
-        } else {
-            alert("Кажется, у вас сломался интернет, попробуйте позже");
-        };
-    })
-};
+// function getFromApi() {
+//     return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
+//         method: "GET",
+//     }).then((responce) => {
+//         if (responce.status === 500) {
+//             throw new Error("Ошибка 500");
+//         } else {
+//             return responce.json();
+//         }
+//     }).then((responceData) => {
+//         comments = responceData.comments;
+//         renderComments(listElem, getCommentsList);
+//     }).catch((error) => {
+//         if (error.message === "Ошибка 500") {
+//             console.log(error);
+//             console.log("Сервер сломался, попробуй позже");
+//         } else {
+//             console.log("Кажется, у вас сломался интернет, попробуйте позже");
+//         };
+//     })
+// };
 
 
 function delay(interval = 300) {
@@ -103,7 +93,7 @@ const likeButtonsListeners = () => {
             delay(2000).then(() => {
                 comment.likes = comment.isLiked ? comment.likes - 1 : comment.likes + 1;
                 comment.isLiked = !comment.isLiked;
-                renderComments();
+                renderComments(listElem, getCommentsList);
             });
         });
     };
@@ -129,7 +119,7 @@ const editButtonsListeners = () => {
             comment.isEdit = !comment.isEdit;
             comments[index].isEdit = comment.isEdit;
 
-            renderComments();
+            renderComments(listElem, getCommentsList);
         });
     };
 };
@@ -138,85 +128,23 @@ const editButtonsListeners = () => {
 const replyToCommentListeners = () => {
     const commentTextElements = document.querySelectorAll('.comment');
     for (const commentText of commentTextElements) {
-        commentText.addEventListener('click', () => {
+        commentText.addEventListener('click', (event) => {
             event.stopPropagation();
             const index = commentText.dataset.index;
             const comment = comments[index];
             addComment.classList.add('reply-comment');
             addComment.value = `QUOTE_BEGIN ${comment.name}:\n${comment.text.replaceAll("QUOTE_BEGIN", "").replaceAll("QUOTE_END", "")}QUOTE_END\n `;
             addComment.focus();
-            renderComments();
+            renderComments(listElem, getCommentsList);
         });
     };
 };
 
-
-const renderComments = () => {
-    const commentsHtml = comments
-        .map((comment, index) => {
-            return `<li class="comment" data-index="${index}">
-          <div class="comment-header">
-            <div>
-              ${comment.author.name}
-            </div>
-            <div>
-              ${getActualDate(comment.date)}
-            </div>
-          </div>
-          <div class="comment-body">
-            ${comment.isEdit
-                    ? `<textarea type="textarea" class="comment-text" rows="4" onclick="event.stopPropagation()">${comment.text}</textarea>`
-                    : `<div class="comment-text" data-index="${index}">
-              ${comment.text.replaceAll("QUOTE_BEGIN", "<div class='quote'>").replaceAll("QUOTE_END", "</div>")}
-            </div>`
-                }
-          </div>
-          <div class="comment-footer">
-            <div>
-              <button class="add-form-button edit-button" data-index="${index}">${comment.isEdit ? 'Сохранить' : 'Редактировать'
-                }</button>
-            </div>
-            <div class="likes">
-              <span class="likes-counter">
-                ${comment.likes}
-              </span>
-              <button class="like-button ${comment.isLiked ? '-active-like' : ''}" data-index="${index}">
-              </button>
-            </div>
-          </div>
-        </li>`;
-        })
-        .join('');
-
-    listElem.innerHTML = commentsHtml;
-    likeButtonsListeners();
-    editButtonsListeners();
-    replyToCommentListeners();
-};
-
-renderComments();
+export { likeButtonsListeners, editButtonsListeners, replyToCommentListeners };
 
 
-function getActualDate(date) {
-    let userDate = new Date(date);
-    let month = userDate.getMonth() + 1;
-    let minutes = userDate.getMinutes();
-    let year = userDate.getFullYear() % 100;
-
-    if (month < 10) {
-        month = "0" + month;
-    };
-
-    if (minutes < 10) {
-        minutes = "0" + minutes;
-    };
-
-    if (year < 10) {
-        year = "0" + year;
-    };
-
-    return `${userDate.getDate()}.${month}.${year} ${userDate.getHours()}:${minutes}`;
-};
+renderComments(listElem, getCommentsList);
+console.log(renderComments(listElem, getCommentsList));
 
 
 const addNewComment = () => {
@@ -255,50 +183,50 @@ const addNewComment = () => {
         };
 
 
-        function postToApi() {
-            addFormElem.style.display = 'none';
-            commentAddedElem.style.display = 'block';
+        // function postToApi() {
+        //     addFormElem.style.display = 'none';
+        //     commentAddedElem.style.display = 'block';
 
-            return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
-                method: "POST",
-                body: JSON.stringify({
-                    text: escapeHtml(addComment.value.trim()),
-                    name: escapeHtml(addName.value.trim()),
-                    forceError: true,
-                })
-            }).then((responce) => {
-                if (responce.status === 400) {
-                    throw new Error("Ошибка 400");
-                } else if (responce.status === 500) {
-                    throw new Error("Ошибка 500");
-                } else {
-                    console.log(responce.status);
-                    return responce.json();
-                }
-            }).then((responceData) => {
-                return getFromApi();
-            }).then(() => {
-                changeMessageToAddForm();
-                addName.value = '';
-                addComment.value = '';
-            }).catch((error) => {
-                if (error.message === "Ошибка 400") {
-                    console.log(error);
-                    changeMessageToAddForm();
-                    alert("Имя и комментарий должны быть не короче 3 символов");
-                } else if (error.message === "Ошибка 500") {
-                    console.log(error);
-                    changeMessageToAddForm();
-                    // alert("Сервер сломался, попробуй позже");
-                    postToApi();
-                } else {
-                    changeMessageToAddForm();
-                    alert("Кажется, у вас сломался интернет, попробуйте позже");
-                };
-            })
-        };
+        //     return fetch("https://webdev-hw-api.vercel.app/api/v1/freddy-krugliy/comments", {
+        //         method: "POST",
+        //         body: JSON.stringify({
+        //             text: escapeHtml(addComment.value.trim()),
+        //             name: escapeHtml(addName.value.trim()),
+        //             forceError: true,
+        //         })
+        //     }).then((responce) => {
+        //         if (responce.status === 400) {
+        //             throw new Error("Ошибка 400");
+        //         } else if (responce.status === 500) {
+        //             throw new Error("Ошибка 500");
+        //         } else {
+        //             console.log(responce.status);
+        //             return responce.json();
+        //         }
+        //     }).then((responceData) => {
+        //         return getFromApi();
+        //     }).then(() => {
+        //         changeMessageToAddForm();
+        //         addName.value = '';
+        //         addComment.value = '';
+        //     }).catch((error) => {
+        //         if (error.message === "Ошибка 400") {
+        //             console.log(error);
+        //             changeMessageToAddForm();
+        //             alert("Имя и комментарий должны быть не короче 3 символов");
+        //         } else if (error.message === "Ошибка 500") {
+        //             console.log(error);
+        //             changeMessageToAddForm();
+        //             // alert("Сервер сломался, попробуй позже");
+        //             postToApi();
+        //         } else {
+        //             changeMessageToAddForm();
+        //             alert("Кажется, у вас сломался интернет, попробуйте позже");
+        //         };
+        //     })
+        // };
 
-        renderComments();
+        renderComments(listElem, getCommentsList);
         postToApi();
     });
 };
@@ -342,7 +270,7 @@ const addNewCommentOnEnter = (event) => {
 const DeleteLastComment = () => {
     lastCommentDeleteButton.addEventListener('click', () => {
         comments.pop();
-        renderComments();
+        renderComments(listElem, getCommentsList);
     });
 };
