+++ index.html
@@ -24,8 +24,7 @@
     </div>
     <div class="comment-added" style="display: none;">Комментарий добавляется...</div>
   </div>
+  
+  <script type="module" src="./JS/main.js"></script>
 </body>
-
-<script type="module" src="./JS/main.js"></script>
-
 </html>
